+++ packages/act_thumbhash/lib/act_thumbhash.dart
@@ -1,3 +1,1 @@
-library act_thumbhash;
-
 export 'src/thumbhash_base.dart';